Functions/model-training-and-test/requirements.txt
functions-framework==3.*
db-dtypes
numpy
pandas
matplotlib
scikit-learn
google-cloud-bigquery
google-cloud-storage
tensorflow
joblib